Move admin features to admin module.
